// taskManagement/src/app/api/model/taskDto.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { TaskDtoStatus } from './taskDtoStatus';
import type { UserDto } from './userDto';

export interface TaskDto {
  id?: number;
  title?: string;
  description?: string;
  status?: TaskDtoStatus;
  projectId?: number;
  users?: UserDto[];
}